Remove polygonOffset from grate textures
This is not a good use of polygonOffset. It is meant for a surface that overlaps another surface but needs to be drawn above it.

The use of polygonOffset here seems to be a trigger for https://github.com/DaemonEngine/Daemon/issues/1676. It may be related to the fact that polygonOffset disables the depth shader from being generated. No depth shader also has the deleterious effect of making realtime lights unable to affect the surface.